pyproject.toml
[project]
name = "holdco"
version = "0.2.0"
description = "Open source holding company management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-unfold>=0.40",
    "pydantic>=2.0",
    "yfinance>=0.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-django>=4.8"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "holdco.settings"
testpaths = ["core/tests"]
